% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litReview-package.R
\docType{package}
\name{litReview-package}
\alias{litReview}
\alias{litReview-package}
\title{litReview: Summarizing Graphs for Literature Reviews}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions to summarize and visualize categorical data from literature reviews. Generates bar charts, waffle charts, donut plots, co-occurrence heatmaps, trend charts, and world maps for reporting study characteristics across multiple publications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sonsoleslp/litReview}
  \item \url{http://sonsoles.me/litReview/}
  \item Report bugs at \url{https://github.com/sonsoleslp/litReview/issues}
}

}
\author{
\strong{Maintainer}: Sonsoles López-Pernas \email{sonsoles.lopez@uef.fi}

Authors:
\itemize{
  \item Kamila Misiejuk
  \item Mohammed Saqr
}

}
\keyword{internal}
