% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{PALETTE}
\alias{PALETTE}
\title{Default Color Palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
PALETTE
}
\description{
A character vector of 8 hex colors for use in literature review plots.
}
\keyword{datasets}
